move the CoS task intake mapping out of addTask and pin it to the request schema
`addTask` in cosTaskStore.js had grown to 400 lines, 250 of which were the
pure mapping from a request's top-level fields onto the persisted `metadata`.
The request schema (`createCosTaskSchema`, lib/cosValidation.js) lived in
another file with nothing checking the two agreed, and they drifted three
times: `optionalReviewers` was validated but never persisted (the task form's
`~opt` badges silently fell back to the Code Review Defaults), and
`discardWorktree` / `noCodeOutput` were reachable only through the raw path.
Over the last 90 days 13 of the 25 feature commits that touched `addTask`
also edited the schema, so every new task option paid that two-file cost.

The mapping now lives in `services/cosTaskIntake.js` as `buildTaskMetadata` /
`buildQueuedTask` (a verbatim move, no behavior change), and
`cosTaskIntake.test.js` classifies every schema field: it either maps (a
schema-valid sample must change the metadata) or is named as a non-metadata
field with the reason. Adding a schema field without either fails the suite;
dropping a mapping fails the field's sample.

`PRIORITY_VALUES` was a second copy of taskParser.js's table; the store now
re-exports the original at the same address.
